+++ packages/core/src/controllers/ApiController.ts
@@ -30,6 +30,7 @@ export interface ApiControllerState {
   wallets: WcWallet[]
   search: WcWallet[]
   isAnalyticsEnabled: boolean
+  excludedRDNS: string[]
 }
 
 type StateKey = keyof ApiControllerState
@@ -42,7 +43,8 @@ const state = proxy<ApiControllerState>({
   recommended: [],
   wallets: [],
   search: [],
-  isAnalyticsEnabled: false
+  isAnalyticsEnabled: false,
+  excludedRDNS: []
 })
 
 // -- Controller ---------------------------------------- //
@@ -61,6 +63,14 @@ export const ApiController = {
       'x-sdk-type': sdkType,
       'x-sdk-version': sdkVersion
     }
+  },
+
+  _filterOutExtensions(wallets: WcWallet[]) {
+    if (OptionsController.state.isUniversalProvider) {
+      return wallets.filter(w => Boolean(w.mobile_link || w.desktop_link || w.webapp_link))
+    }
+
+    return wallets
   },
 
   async _fetchWalletImage(imageId: string) {
@@ -94,7 +104,8 @@ export const ApiController = {
   },
 
   async fetchNetworkImages() {
-    const { requestedCaipNetworks } = NetworkController.state
+    const requestedCaipNetworks = NetworkController.getRequestedCaipNetworks()
+
     const ids = requestedCaipNetworks?.map(({ imageId }) => imageId).filter(Boolean)
     if (ids) {
       await Promise.allSettled((ids as string[]).map(id => ApiController._fetchNetworkImage(id)))
@@ -188,9 +199,34 @@ export const ApiController = {
       ...(images as string[]).map(id => ApiController._fetchWalletImage(id)),
       CoreHelperUtil.wait(300)
     ])
-    state.wallets = [...state.wallets, ...data]
+
+    state.wallets = CoreHelperUtil.uniqueBy(
+      [...state.wallets, ...ApiController._filterOutExtensions(data)],
+      'id'
+    )
     state.count = count > state.count ? count : state.count
     state.page = page
+  },
+
+  async searchWalletByIds({ ids }: { ids: string[] }) {
+    const { data } = await api.get<ApiGetWalletsResponse>({
+      path: '/getWallets',
+      headers: ApiController._getApiHeaders(),
+      params: {
+        page: '1',
+        entries: String(ids.length),
+        chains: NetworkController.state.caipNetwork?.id,
+        include: ids?.join(',')
+      }
+    })
+
+    if (data) {
+      data.forEach(wallet => {
+        if (wallet?.rdns) {
+          state.excludedRDNS.push(wallet.rdns)
+        }
+      })
+    }
   },
 
   async searchWallet({ search }: Pick<ApiGetWalletsRequest, 'search'>) {
@@ -213,7 +249,7 @@ export const ApiController = {
       ...(images as string[]).map(id => ApiController._fetchWalletImage(id)),
       CoreHelperUtil.wait(300)
     ])
-    state.search = data
+    state.search = ApiController._filterOutExtensions(data)
   },
 
   async reFetchWallets() {