+++ packages/scaffold-ui/exports/w3m-modal.ts
@@ -1,0 +1,1 @@
+export * from '../src/modal/w3m-modal/index.js'