+++ packages/scaffold-ui/src/views/w3m-account-view/index.ts
@@ -1,0 +1,34 @@
+import { ConnectorController, StorageUtil } from '@web3modal/core'
+import { customElement } from '@web3modal/ui'
+import { LitElement, html } from 'lit'
+
+@customElement('w3m-account-view')
+export class W3mAccountView extends LitElement {
+  // -- Render -------------------------------------------- //
+
+  public override render() {
+    const type = StorageUtil.getConnectedConnector()
+    const authConnector = ConnectorController.getAuthConnector()
+
+    return html`
+      ${authConnector?.walletFeatures && type === 'AUTH'
+        ? this.walletFeaturesTemplate()
+        : this.defaultTemplate()}
+    `
+  }
+
+  // -- Private ------------------------------------------- //
+  private walletFeaturesTemplate() {
+    return html`<w3m-account-wallet-features-widget></w3m-account-wallet-features-widget>`
+  }
+
+  private defaultTemplate() {
+    return html`<w3m-account-default-widget></w3m-account-default-widget>`
+  }
+}
+
+declare global {
+  interface HTMLElementTagNameMap {
+    'w3m-account-view': W3mAccountView
+  }
+}